requirements.txt
asgiref            ~= 3.7.2
colorama           ~= 0.4.6
Django             ~= 5.0.4
djangorestframework~= 3.15.1
drf-spectacular    ~= 0.27.2
inflection         ~= 0.5.1
packaging          ~= 24.0
pillow             ~= 10.2.0
pip                ~= 24.0
psycopg2           ~= 2.9.9
pypng              ~= 0.20220715.0
pytz               ~= 2024.1
PyYAML             ~= 6.0.1
qrcode             ~= 7.4.2
sqlparse           ~= 0.4.4
typing_extensions  ~= 4.10.0
tzdata             ~= 2024.1
uritemplate        ~= 4.1.1